+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 Change Log
 ==========
 
+0.4.1 (unreleased)
+------------------
+
+- Raise TimeoutErrors when services fail to start from any causing exception.
+- Use more botocore client-based service checks.
+
 0.4.0 (2019-08-21)
 ------------------